Revert "Taxonomy CV mapping, OpenContext setup "
