+++ scripts/taxonomy/classification.py
@@ -1,6 +1,6 @@
 import click
 import click_config_file
-from isamples_metadata import SESARTransformer, OpenContextTransformer
+from isamples_metadata import SESARTransformer
 
 import isb_lib
 from isb_lib.core import ThingRecordIterator
@@ -8,7 +8,6 @@ from isb_web.sqlmodel_database import SQLModelDAO
 from create_hierarchy_json import getFullLabel, getHierarchyMapping
 from classification_helper import classify_by_machine, classify_by_rule
 from SESARClassifierInput import SESARClassifierInput
-from OpenContextClassifierInput import OpenContextClassifierInput
 
 
 def get_classification_result(description_map, text, collection, labelType):
@@ -39,29 +38,21 @@ def get_classification_result(description_map, text, collection, labelType):
     help="Maximum records to load, -1 for all",
 )
 @click.option(
-    "-a",
-    "--authority_id",
-    default="SESAR",
-    help="Collection type",
-)
-@click.option(
     "-v",
     "--verbosity",
     default="DEBUG",
     help="Specify logging level",
     show_default=True,
 )
+# You can specify the filename by doing --config <file> as the cmdline option
 @click_config_file.configuration_option()
 @click.pass_context
-def main(
-    ctx, db_url: str, solr_url: str, max_records: int,
-    authority_id: str, verbosity: str
-):
+def main(ctx, db_url: str, solr_url: str, max_records: int, verbosity: str):
     isb_lib.core.things_main(ctx, db_url, solr_url, verbosity)
     db_session = SQLModelDAO(db_url).get_session()
     thing_iterator = ThingRecordIterator(
         db_session,
-        authority_id=authority_id,
+        authority_id="SESAR",
         page_size=max_records
     )
     context_mapping = getHierarchyMapping("context")
@@ -70,62 +61,38 @@ def main(
 
     for thing in thing_iterator.yieldRecordsByPage():
         # print(f"thing is {thing.id}")
-        if authority_id == "SESAR":
-            transformed = SESARTransformer.SESARTransformer(
-                thing.resolved_content
-            ).transform()
+        transformed = SESARTransformer.SESARTransformer(
+            thing.resolved_content
+        ).transform()
 
-            # parse the thing to classifier input form
-            input = SESARClassifierInput(thing.resolved_content)
-            input.parse_thing()
+        # parse the thing to classifier input form
+        parsed = SESARClassifierInput(thing.resolved_content).parse_thing()
 
-            description_map = input.get_description_map()
-            material_text = input.get_material_text()
+        description_map = parsed.get_description_map()
+        material_text = parsed.get_material_text()
 
-            # get the material label prediction result of the record
-            label, prob = get_classification_result(
-                description_map, material_text, "SESAR", "material"
-            )
+        # get the material label prediction result of the record
+        label, prob = get_classification_result(
+            description_map, material_text, "SESAR", "material"
+        )
 
-            print(
-                f"Predicted (probability, label) : {label}, {prob}"
-            )
+        print(
+            f"Predicted (probability, label) : {label}, {prob}"
+        )
 
-            # gold label of the record
-            context_label = transformed['hasContextCategory']
-            material_label = transformed['hasMaterialCategory']
-            specimen_label = transformed['hasSpecimenCategory']
+        # gold label of the record
+        context_label = transformed['hasContextCategory']
+        material_label = transformed['hasMaterialCategory']
+        specimen_label = transformed['hasSpecimenCategory']
 
-            full_context = [getFullLabel(context_label, context_mapping)]
-            full_material = [getFullLabel(material_label, material_mapping)]
-            full_specimen = [getFullLabel(specimen_label, specimen_mapping)]
+        full_context = [getFullLabel(context_label, context_mapping)]
+        full_material = [getFullLabel(material_label, material_mapping)]
+        full_specimen = [getFullLabel(specimen_label, specimen_mapping)]
 
-            # print out the full hierarchy of the label
-            print(
-                f"Gold label (hierarchical) :  "
-                f"material: {full_material} / ",
-                f"specimen: {full_specimen} / "
-                f"context: {full_context} ",
-            )
-
-            print("-----------------")
-        else:
-
-            transformed = OpenContextTransformer.OpenContextTransformer(
-                thing.resolved_content
-            ).transform()
-
-            # parse the thing to classifier input form
-            input = OpenContextClassifierInput(thing.resolved_content)
-            input.parse_thing()
-
-            material_text = input.get_material_text()
-            sample_text = input.get_sample_text()
-
-            # print out the classifier input form of this record
-            print(material_text)
-            print(sample_text)
-
+        # print out the full hierarchy of the label
+        print(f"context: {full_context} material: {full_material} "
+              f"specimen: {full_specimen}")
+        print("-----------------")
     db_session.close()
